requirements.txt
selenium
webdriver-manager
twilio
python-dotenv
pytest